// src/typings/auth/IUser.ts
export interface IUser {
    id?: number;
    username: string;
    password: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    gender?: string;
    mobilePhone?: number;
    dob?: string;
    location?: string;
    facebookAccount?: string;
    googleAccount?: string;
    twitterAccount?: string;
    logo?: string;
}